# Passthrough non-transformed strings in compressWhitespace