After
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/transformations
BenchmarkCompressWhitespace
BenchmarkCompressWhitespace/#00
BenchmarkCompressWhitespace/#00-10         	571122726	         2.125 ns/op
BenchmarkCompressWhitespace/test
BenchmarkCompressWhitespace/test-10        	248496358	         4.828 ns/op
BenchmarkCompressWhitespace/test_case
BenchmarkCompressWhitespace/test_case-10   	29280343	        40.49 ns/op
BenchmarkCompressWhitespace/test____case
BenchmarkCompressWhitespace/test____case-10         	25941014	        46.05 ns/op
BenchmarkCompressWhitespace/_test__c__as___
BenchmarkCompressWhitespace/_test__c__as___-10      	17068617	        70.48 ns/op
PASS
```

Before
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/transformations
BenchmarkCompressWhitespace
BenchmarkCompressWhitespace/#00
BenchmarkCompressWhitespace/#00-10         	447815151	         2.590 ns/op
BenchmarkCompressWhitespace/test
BenchmarkCompressWhitespace/test-10        	32739363	        36.20 ns/op
BenchmarkCompressWhitespace/test_case
BenchmarkCompressWhitespace/test_case-10   	16433900	        72.80 ns/op
BenchmarkCompressWhitespace/test____case
BenchmarkCompressWhitespace/test____case-10         	14380442	        83.88 ns/op
BenchmarkCompressWhitespace/_test__c__as___
BenchmarkCompressWhitespace/_test__c__as___-10      	12626217	        96.18 ns/op
PASS
```